% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redalyc_add_affiliations_c1.R
\name{redalyc_add_affiliations_c1}
\alias{redalyc_add_affiliations_c1}
\title{Construye Affiliations y C1 estilo Scopus/bibliometrix desde Redalyc}
\usage{
redalyc_add_affiliations_c1(
  df,
  authors_delim = ",",
  inst_delim = " ",
  aff_sep = ";",
  upper = TRUE
)
}
\arguments{
\item{df}{data.frame/tibble con: autores, institucionAutores,
institucionRevista, nomInstitucionRev}

\item{authors_delim}{separador de autores (en tu caso coma)}

\item{inst_delim}{separador de ids de institución del autor (en tu caso espacio)}

\item{aff_sep}{separador final entre afiliaciones/autores (bibliometrix usa ";")}

\item{upper}{logical: convertir a mayúsculas}
}
\value{
df con dos columnas nuevas: Affiliations y C1
}
\description{
Construye Affiliations y C1 estilo Scopus/bibliometrix desde Redalyc
}
