% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redalyc_get_articles.R
\name{redalyc_get_articles}
\alias{redalyc_get_articles}
\title{Obtiene metadatos de artículos desde Redalyc (getArticles)}
\usage{
redalyc_get_articles(query, page = 1, page_size = 100, ...)
}
\arguments{
\item{query}{Texto de búsqueda.}

\item{page}{Página a consultar.}

\item{page_size}{Registros por página.}

\item{...}{Parámetros adicionales pasados a redalyc_build_url (collection, sort, base_url).}
}
\value{
Lista con totalResultados y resultados (tibble).
}
\description{
Obtiene metadatos de artículos desde Redalyc (getArticles)
}
