% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redalyc_write_bib.R
\name{redalyc_write_bib}
\alias{redalyc_write_bib}
\title{Exportar a BibTeX (artículos) desde data frame bibliometrix-like}
\usage{
redalyc_write_bib(M, path, key_strategy = c("author_year_title", "ut"))
}
\arguments{
\item{M}{data.frame bibliometrix-like (idealmente salida de redalyc_as_bibliometrix()).}

\item{path}{Ruta destino .bib}

\item{key_strategy}{"ut" o "author_year_title"}
}
\description{
Exportar a BibTeX (artículos) desde data frame bibliometrix-like
}
