% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redalycR_output_dir.R
\name{redalycR_output_dir}
\alias{redalycR_output_dir}
\title{Get redalycR output directory}
\usage{
redalycR_output_dir(create = FALSE)
}
\arguments{
\item{create}{Logical; if \code{TRUE}, the directory is created if it does not exist.}
}
\value{
A character string with the output directory path.
}
\description{
Returns the directory used by redalycR to save plots and tables. By default,
it uses \code{~/Documents/output_redalycR}. Users can override this path via
\code{options(redalycR.output_dir = "...")}.
}
\examples{
redalycR_output_dir()

}
