% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redalyc_as_bibliometrics.R
\name{redalyc_as_bibliometrics}
\alias{redalyc_as_bibliometrics}
\title{Convertir resultados Redalyc a estructura bibliometrix (Field Tags)}
\usage{
redalyc_as_bibliometrics(x, sep = ";", db = "REDALYC")
}
\arguments{
\item{x}{data.frame/tibble devuelto por redalyc_get_all_articles() o redalyc_get_articles()$resultados}

\item{sep}{Separador estándar para campos multivaluados (bibliometrix usa ";" por defecto).}

\item{db}{Nombre de base para DB (por defecto "REDALYC").}
}
\value{
tibble con columnas tipo bibliometrix (AU, TI, SO, DE, ID, AB, PY, DI, UT, DB, ...).
}
\description{
Convertir resultados Redalyc a estructura bibliometrix (Field Tags)
}
